tjw-28: WIP

+++ scripts/setup.sh
@@ -24,15 +24,19 @@ command_exists() {
 
 # Function to install Homebrew if missing
 install_homebrew() {
-    echo "Homebrew is not installed."
-    echo "Installing Homebrew..."
-    /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)" || { echo "Failed to install Homebrew."; exit 1; }
-    
-    # Add Homebrew to PATH for the current session
-    if [[ -x /opt/homebrew/bin/brew ]]; then
-        eval "$(/opt/homebrew/bin/brew shellenv)"
-    elif [[ -x /usr/local/bin/brew ]]; then
-        eval "$(/usr/local/bin/brew shellenv)"
+    if ! command_exists brew; then
+        echo "Homebrew is not installed."
+        echo "Installing Homebrew..."
+        /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)" || { echo "Failed to install Homebrew."; exit 1; }
+        
+        # Add Homebrew to PATH for the current session
+        if [[ -x /opt/homebrew/bin/brew ]]; then
+            eval "$(/opt/homebrew/bin/brew shellenv)"
+        elif [[ -x /usr/local/bin/brew ]]; then
+            eval "$(/usr/local/bin/brew shellenv)"
+        fi
+    else
+        debug "Homebrew is already installed"
     fi
 }
 
@@ -42,6 +46,7 @@ install_whiptail() {
         echo "whiptail is not installed."
         
         if [[ "$PLATFORM" == "MacOS" ]]; then
+            install_homebrew
             echo "Installing whiptail via Homebrew..."
             brew install newt || { echo "Failed to install whiptail."; exit 1; }
         elif [[ "$PLATFORM" == "Ubuntu" ]]; then
@@ -62,6 +67,7 @@ install_ansible() {
         echo "Ansible is not installed."
         
         if [[ "$PLATFORM" == "MacOS" ]]; then
+            install_homebrew
             echo "Installing Ansible via Homebrew..."
             brew install ansible || { echo "Failed to install Ansible."; exit 1; }
         elif [[ "$PLATFORM" == "Ubuntu" ]]; then
